리팩토링: PnpmRunner 서비스 추출
- Editor/Package/PnpmRunner.cs 신설 (pnpm install sync/async/background)
- facade.StartPnpmInstallInBackground는 1줄 위임 유지 (AITConvertCore 외부 호환)
- RunPnpmInstallInThread/Sync/Async는 facade에서 제거
- InstallStages는 PnpmStoreManager, CleanNodeModules는 NodeModulesValidator 사용
- PackageContext를 internal로 승격 (서비스 접근용)
- 테스트: 프로세스 실행 위주이므로 E2E에 위임
